move pause_manager to it's own file
