+++ api/src/opentrons/hardware_control/types.py
@@ -106,36 +106,3 @@ class MustHomeError(RuntimeError):
 
 class NoTipAttachedError(RuntimeError):
     pass
-
-def pausable(func):
-    @functools.wraps(func)
-    async def pausable_wrapped_obj(*args, **kwargs):
-        self = args[0]
-        try:
-            await self.pause_manager.hold_while_paused()
-        except AttributeError as e:
-            MODULE_LOG.exception(f'Pausable cannot decorate function without'
-                                 f'reference to pause_manager: {e}')
-        return func(*args, **kwargs)
-    return pausable_wrapped_obj
-
-
-class PauseManager():
-    def __init__(self, loop: asyncio.AbstractEventLoop, is_simulating: bool):
-        self._run_flag = asyncio.Event(loop=loop)
-        self._is_simulating = is_simulating
-
-    async def hold_while_paused(self):
-        if self._is_simulating:
-            return True
-        else:
-            return await self._run_flag.wait()
-
-    def pause():
-        self._run_flag.clear()
-
-    def resume():
-        self._run_flag.set()
-
-    def is_paused():
-        self._run_flag.is_set()
