+++ api/src/opentrons/protocol_api/util.py
@@ -208,7 +208,7 @@ class HardwareManager:
         return self._current
 
     def set_hw(self, hardware):
-        if self._current and (self._is_orig or self._built_own_adapter):
+        # if self._current and (self._is_orig or self._built_own_adapter):
             # self._current.join()
         MODULE_LOG.info(f'set_hw hardware: {hardware}')
         MODULE_LOG.info(f'set_hw api: {hardware._api}')
